pulsewise-be/requirements.txt
fastapi[standard]
uvicorn
SQLAlchemy
psycopg[binary]
alembic
pydantic-settings
passlib[bcrypt]
python-jose[cryptography]
numpy
